Move aggregator database::provider::test_helper to its parent database module
Since it may be used in multiple database submodules now.
