+++ mithril-aggregator/src/database/provider/mod.rs
@@ -8,8 +8,6 @@ mod signer;
 mod signer_registration;
 mod single_signature;
 mod stake_pool;
-#[cfg(test)]
-mod test_helper;
 
 pub use cardano_transaction::*;
 pub use certificate::*;
@@ -20,5 +18,3 @@ pub use signer::*;
 pub use signer_registration::*;
 pub use single_signature::*;
 pub use stake_pool::*;
-#[cfg(test)]
-pub use test_helper::*;
