+++ mithril-aggregator/src/database/provider/signed_entity.rs
@@ -390,7 +390,7 @@ mod tests {
     };
     use mithril_common::{entities::CardanoDbBeacon, test_utils::fake_data};
 
-    use crate::database::provider::{apply_all_migrations_to_db, disable_foreign_key_support};
+    use crate::database::test_helper::{apply_all_migrations_to_db, disable_foreign_key_support};
 
     use super::*;
