+++ mithril-aggregator/src/database/provider/stake_pool.rs
@@ -234,7 +234,7 @@ impl StakeStorer for StakePoolStore {
 mod tests {
     use sqlite::Connection;
 
-    use crate::database::provider::apply_all_migrations_to_db;
+    use crate::database::test_helper::apply_all_migrations_to_db;
 
     use super::*;
