+++ mithril-aggregator/src/tools/signer_importer.rs
@@ -190,9 +190,8 @@ mod tests {
 
     use warp::Filter;
 
-    use crate::database::provider::{
-        apply_all_migrations_to_db, disable_foreign_key_support, SignerGetter, SignerStore,
-    };
+    use crate::database::provider::{SignerGetter, SignerStore};
+    use crate::database::test_helper::{apply_all_migrations_to_db, disable_foreign_key_support};
     use crate::http_server::routes::reply;
 
     use super::*;
